requirements.txt
Django>=5.2.4
djangorestframework>=3.15.1
pytest>=8.4.1
pytest-django>=4.11.1 